chore(prerelease): move topbar to component, and inside titlebar instead of same levela
