+++ ui/src/components/topbar/mod.rs
@@ -1,0 +1,1 @@
+pub mod release_info;
